Fix #24
This is a long-standing issue. Discord cuts of a message on the 200th spoiler tag or emoji, and because this bot uses a lot of both, it often causes games to be incomplete. This fix makes the bot split up a game into multiple messages more aggressively. As far as I can tell, this issue never occurs anymore.