Implement #33
This new '/auto' command can only be run by members with Manage Channels permissions, and it can make the bot start sending Minesweeper games at regular intervals in a chosen channel. The data is saved in an autoChannels.json file.

This commit also adds some properties to settings.json: the default game settings, and the maximum interval for the auto command.

Furthermore, the logic for checking game settings and creating the game has now been split into separate functions in generateGame.js, and game settings are now one object instead of separate parameters.

Lastly, a command run function can now return a Promise and it will be awaited before replying.

Some of the code in this commit is based on the fork INotSleep/minesweeper-bot

Co-authored-by: INotSleep <77581131+INotSleep@users.noreply.github.com>